Dodani underflow i overflow protection. Tau dis estimator je sumnjiv. Potrebno odrediti vrijednost kontsanti.

+++ drivers/inc/lp_filter.h
@@ -2,11 +2,11 @@
 #define LOW_PASS_FILTER_H_
 
 typedef struct {
-    signed long alpha;
+    Uint32 alpha;
     signed long y_prev;
 } LowPassFilter;
 
-void filter_init(LowPassFilter *filter, int alpha, int y0);
+void filter_init(LowPassFilter *filter, Uint32 alpha, signed long y0);
 signed long apply_filter(LowPassFilter *filter, signed long x);
 
 #endif /* LOW_PASS_FILTER_H_ */
